requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-chess==1.999
gevent==23.9.1
gevent-websocket==0.10.1
gunicorn==21.2.0
